+++ guillotina/contrib/dbusers/content/groups.py
@@ -6,8 +6,6 @@ from guillotina.directives import index_field
 from guillotina.interfaces import IFolder
 from zope.interface import implementer
 
-import typing
-
 
 class IGroupManager(IFolder):
     pass
@@ -33,9 +31,11 @@ class IGroup(IFolder):
     behaviors=["guillotina.behaviors.dublincore.IDublinCore"],
 )
 class Group(Folder):
-    name = description = None
-    user_roles: typing.List[str] = []
-    users: typing.List[str] = []
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.name = self.description = None
+        self.user_roles = []
+        self.users = []
 
     @property
     def roles(self):